Display comment when the popup loads
