+++ src/assets/stylesheets/style.css
@@ -159,15 +159,9 @@ footer div {
   column-gap: 1rem;
 }
 
-.ul-list {
-  margin-top: 2rem;
-  display: flex;
-  flex-direction: column;
-  row-gap: 1rem;
-}
-
 .list-item {
   margin-bottom: 1rem;
+  text-align: center;
 }
 
 .summary {
