--- src/modules/getLikes.js
@@ -1,6 +1,0 @@
-const getLikes = async () => {
-  const response = await fetch('https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/IEprda6bVbN0jOFjgzgK/likes')
-  return response.json()
-}
-
-export default getLikes
